refactor: Move function not related to data retrieval but to data refcatoring in utils.py
